requirements.txt
PyYAML>=6.0.1
discord>=2.3.2
yt-dlp>=2023.12.30
pydantic==2.5.3